// src/archive/automation/actions/types.ts
export interface ActionResult {
  success: boolean;
  message: string;
}